requirements.txt
fastapi==0.89.1
numpy==1.21.3
openai==0.19.0
opencv_python==4.5.1.48
Pillow==9.4.0
pydantic==1.9.0
pytesseract==0.3.8
starlette==0.16.0
